requirements.txt
mysql-connector-python==9.3.0 
streamlit
streamlit-authenticator ==0.4.2